chore: try to add decimal type

+++ bc_obps/registration/api.py
@@ -7,6 +7,7 @@ from django.shortcuts import get_object_or_404
 from .models import Operation, Operator, NaicsCode
 from ninja.orm import create_schema
 from ninja import Field, Schema
+import decimal
 
 
 router = Router()
@@ -38,13 +39,13 @@ class OperationIn(Schema):
     ghfrp_id: str
     bcghrp_id: str
     petrinex_id: str
-    latitude: int
-    longitude: int
+    latitude: decimal
+    longitude: decimal
     legal_land_description: str
     nearest_municipality: str
-    operator_percent_of_ownership: int
+    operator_percent_of_ownership: decimal
     registered_for_obps: bool
-    estimated_emissions: int
+    estimated_emissions: decimal
     registered_for_obps: str = Field(default=False)
     # contacts:
     # documents:
@@ -62,13 +63,13 @@ class OperationOut(Schema):
     ghfrp_id: str
     bcghrp_id: str
     petrinex_id: str
-    latitude: int
-    longitude: int
+    latitude: decimal
+    longitude: decimal
     legal_land_description: str
     nearest_municipality: str
-    operator_percent_of_ownership: int
+    operator_percent_of_ownership: decimal
     registered_for_obps: bool
-    estimated_emissions: int
+    estimated_emissions: decimal
     # contacts:
     # documents:
